fixed numpy import errors
