+++ sympy/plotting/intervalmath/tests/test_interval_functions.py
@@ -1,3 +1,4 @@
+from __future__ import division
 from sympy.plotting.intervalmath import *
 from sympy.external import import_module
 
@@ -5,7 +6,44 @@ np = import_module('numpy')
 if not np:
     disabled = True
 
+#requires Numpy. Hence included in interval_functions
+def test_interval_pow():
+    a = 2**interval(1, 2) == interval(2, 4)
+    assert a == (True, True)
+    a = interval(1, 2)**interval(1, 2) == interval(1, 4)
+    assert a == (True, True)
+    a = interval(-1, 1)**interval(0.5, 2)
+    assert a.is_valid == None
+    a = interval(-2, -1) ** interval(1, 2)
+    assert a.is_valid == False
+    a = interval(-2, -1) ** (1 / 2)
+    assert a.is_valid == False
+    a = interval(-1, 1)**(1 / 2)
+    assert a.is_valid == None
+    a = interval(-1, 1)**(1 / 3) == interval(-1, 1)
+    assert a == (True, True)
+    a = interval(-1, 1)**2 == interval(0, 1)
+    assert a == (True, True)
+    a = interval(-1, 1) ** (1 / 29) == interval(-1, 1)
+    assert a == (True, True)
+    a = -2**interval(1, 1) == interval(-2, -2)
+    assert a == (True, True)
 
+    a = interval(1, 2, is_valid=False)**2
+    assert a.is_valid == False
+
+    a = (-3)**interval(1, 2)
+    assert a.is_valid == False
+    a = (-4)**interval(0.5, 0.5)
+    assert a.is_valid == False
+    assert ((-3)**interval(1, 1) == interval(-3, -3)) == (True, True)
+
+    a = interval(8, 64)**(2 / 3)
+    assert abs(a.start - 4) < 1e-10 #eps
+    assert abs(a.end - 16) < 1e-10
+    a = interval(-8, 64)**(2 / 3)
+    assert abs(a.start - 4) < 1e-10 #eps
+    assert abs(a.end - 16) < 1e-10
 def test_exp():
     a = exp(interval(-np.inf, 0))
     assert a.start == np.exp(-np.inf)
