+++ sympy/plotting/intervalmath/tests/test_intervalmath.py
@@ -1,7 +1,7 @@
 from __future__ import division
 from sympy.plotting.intervalmath import interval
 from sympy.utilities.pytest import raises
-
+from sympy.external import import_module
 def test_interval():
     assert (interval(1, 1) == interval(1, 1, is_valid=True)) == (True, True)
     assert (interval(1, 1) == interval(1, 1, is_valid=False)) == (True, False)
@@ -186,41 +186,3 @@ def test_interval_div():
     assert a == (True, True)
     a = interval(-5, 5, is_valid=False) / 2
     assert a.is_valid == False
-
-def test_interval_pow():
-    a = 2**interval(1, 2) == interval(2, 4)
-    assert a == (True, True)
-    a = interval(1, 2)**interval(1, 2) == interval(1, 4)
-    assert a == (True, True)
-    a = interval(-1, 1)**interval(0.5, 2)
-    assert a.is_valid == None
-    a = interval(-2, -1) ** interval(1, 2)
-    assert a.is_valid == False
-    a = interval(-2, -1) ** (1 / 2)
-    assert a.is_valid == False
-    a = interval(-1, 1)**(1 / 2)
-    assert a.is_valid == None
-    a = interval(-1, 1)**(1 / 3) == interval(-1, 1)
-    assert a == (True, True)
-    a = interval(-1, 1)**2 == interval(0, 1)
-    assert a == (True, True)
-    a = interval(-1, 1) ** (1 / 29) == interval(-1, 1)
-    assert a == (True, True)
-    a = -2**interval(1, 1) == interval(-2, -2)
-    assert a == (True, True)
-
-    a = interval(1, 2, is_valid=False)**2
-    assert a.is_valid == False
-
-    a = (-3)**interval(1, 2)
-    assert a.is_valid == False
-    a = (-4)**interval(0.5, 0.5)
-    assert a.is_valid == False
-    assert ((-3)**interval(1, 1) == interval(-3, -3)) == (True, True)
-
-    a = interval(8, 64)**(2 / 3)
-    assert abs(a.start - 4) < 1e-10 #eps
-    assert abs(a.end - 16) < 1e-10
-    a = interval(-8, 64)**(2 / 3)
-    assert abs(a.start - 4) < 1e-10 #eps
-    assert abs(a.end - 16) < 1e-10
